+++ monix-reactive/shared/src/main/scala/monix/reactive/internal/consumers/MapConsumer.scala
@@ -1,0 +1,60 @@
+/*
+ * Copyright (c) 2014-2017 by its authors. Some rights reserved.
+ * See the project homepage at: https://monix.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package monix.reactive.internal.consumers
+
+import monix.eval.Callback
+import monix.execution.Scheduler
+import monix.execution.cancelables.AssignableCancelable
+import monix.execution.misc.NonFatal
+import monix.reactive.Consumer
+import monix.reactive.observers.Subscriber
+
+/** Implementation for [[monix.reactive.Consumer.map]]. */
+private[reactive]
+final class MapConsumer[In, R, R2](source: Consumer[In,R], f: R => R2)
+  extends Consumer[In, R2] {
+
+  def createSubscriber(cb: Callback[R2], s: Scheduler): (Subscriber[In], AssignableCancelable) = {
+    val cb1 = new Callback[R] {
+      def onSuccess(value: R): Unit =
+        s.execute(new Runnable {
+          // Forcing an asynchronous boundary, otherwise
+          // this isn't a safe operation.
+          def run(): Unit = {
+            var streamErrors = true
+            try {
+              val r2 = f(value)
+              streamErrors = false
+              cb.onSuccess(r2)
+            } catch {
+              case NonFatal(ex) =>
+                if (streamErrors) cb.onError(ex)
+                else s.reportFailure(ex)
+            }
+          }
+        })
+
+      def onError(ex: Throwable): Unit =
+        s.execute(new Runnable {
+          def run(): Unit = cb.onError(ex)
+        })
+    }
+
+    source.createSubscriber(cb1, s)
+  }
+}
